% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{extract_feed}
\alias{extract_feed}
\title{Extract GitHub RSS feed}
\usage{
extract_feed(feed)
}
\arguments{
\item{feed}{a list containing two elements: a username and a RSS atom link}
}
\value{
a data.frame of the RSS feed contents
}
\description{
Extract GitHub RSS feed
}
