% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{obtain_feed}
\alias{obtain_feed}
\title{Obtain RSS feed URL}
\usage{
obtain_feed(username, password)
}
\arguments{
\item{username}{GitHub username. If missing, will attempt to use GITHUB_PAT to determine it.}

\item{password}{GitHub password. If missing, will be prompted to enter is securely.}
}
\value{
a two-element list containing the username and RSS atom feed URL
}
\description{
Obtain RSS feed URL
}
