% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{starryeyes}
\alias{starryeyes}
\title{Create a shinydashboard of the recent activity}
\usage{
starryeyes(username, password, feedURL)
}
\arguments{
\item{username}{GitHub username. If missing, will attempt to use GITHUB_PAT to determine it.}

\item{password}{GitHub password. If missing, will be prompted to enter is securely.}

\item{feed}{(optionally) a GitHub RSS feed URL (with token) to use}
}
\value{
produces a shinydashboard of recent activity.
}
\description{
Create a shinydashboard of the recent activity
}
